refactor(grammargen): clean up benchmarks, improve grammar extraction and LR splitting
- Remove verbose runtime audit logging and survivor stats from big3 benchmarks to reduce noise
- Fix grammar extraction to infer ProductionIDCount from array dimensions when macro is missing
- Add binary repeat mode support in grammargen for validated grammars (go, graphql, json, etc.)
- Improve LR state splitting with proper lookahead inheritance from LALR follow sets
- Add immediate token support in lexer with AcceptPriority tracking
- Clean up dead code: remove duplicate clearDirtyPathToRoot, unused parser hints, stale tests
- Remove temporary z_tmp_inline_expansion_test.go diagnostic file
- Fix Python incremental test expectation (external scanner unsupported)
- Add diagnostic tooling for Scala block comment lexing investigation

+++ cmd/ts2go/embedded.go
@@ -232,7 +232,7 @@ func convertLexStates(src []LexStateEntry) []gotreesitter.LexState {
 			AcceptToken: gotreesitter.Symbol(s.Accept),
 			Skip:        false,
 			Default:     -1,
-			EOF:         int32(s.EOF),
+			EOF:         s.EOF,
 		}
 		if len(s.Transitions) > 0 {
 			state.Transitions = make([]gotreesitter.LexTransition, len(s.Transitions))
@@ -240,7 +240,7 @@ func convertLexStates(src []LexStateEntry) []gotreesitter.LexState {
 				state.Transitions[j] = gotreesitter.LexTransition{
 					Lo:        t.Lo,
 					Hi:        t.Hi,
-					NextState: int32(t.Next),
+					NextState: t.Next,
 					Skip:      t.Skip,
 				}
 			}
